separates xml specific module from helpers
changes module name from Common to CommonXml to
articulate that these are reinventions of the common.xsl
rules and are specific to XML only! Therefore not including
in generic requirer.rb
renamed test file, four tests fail which was true previous to this commit
