+++ test/classes.rb
@@ -1,1 +1,2 @@
 require_relative "../scripts/ruby/lib/requirer.rb"
+require_relative "../scripts/ruby/lib/common_xml.rb"
